Seemingly working sub-unit panel dupe.
I was having problems with the duplicated sub panel modifying the
selected/un-selected status of the original panel. I think that doing
a deep clone brought over things like the 'element' selectable data
value which points to the element to be modified.
What seems to be working is not doing a deep clone but instead just
expanding the selectable to include these and have it handle adding
the correct selectables data.