Major work to speed up allows/requires checking. Delighted.
A simple permutations check for 13 allows arrays and 9 requires means
a total possible 259 million permutations.
This takes too long in a browser and the browser will think the script
has hung.

So before we do the brute force check I first loop through the
following simplification approaches. Repeating until a complete loop
makes no difference.

Remove any allows sub-elements that are not in the requires.
    This also then allows us to reduce multiple different allows to
    be the same in the case that they only differ by elements that
    aren't required.

Now that we have reduced allows to lowest common denominator we count
all the requires for a particular allow type. If <= to the number of
allows then remove the allows and the same number of matching requires.

Then if we have any allows arrays of length 1 (due to removing
sub-elements not required) then remove the length 1 array and matching
requires - as no other requires can be for this length 1 array.

Then repeat as each step may make a previous step work again.

I'm no longer hitting the recursive permutations check even for
massively complicated situations. I guess that it no longer can but I'm
leaving it in there as a fallback position just in case I'm wrong or
in case any future lists somehow break this.

NB: for this to work all allows arrays must have their elements in the
same order (ascending or descending). I don't bother to sort the
sub-elements as I control their order when writing them. If this pattern
is not followed then lowest common denominator array matching code won't
work, ie [8,9] is not the same as [9,8] even though they are for the
purposes we require. So maintain ordering.